ConstructForest

Description

The cubeForest function constructs a forest in the following manner:

Start with a list of corners all embedded at the origin.

Let each corner be a root of a tree.

For a given corner, pick three edges and embed them in the three spots adjacent to the origin in such a way that no two adjacent stickers have the same colour.

Take such a triplet of edges to be a leaf of the corresponding tree.

Continue this contstruction iteratively: For each leaf, construct a leaf by choosing a set of adjacent pieces, in such a way that no two adjacent stickers have the same colour.

The leaves in such forest define all the cubes with the property that no two adjacent stickers have the same colour.

Documentation

type LayerSeed = (Layer, Set Cubie, Set Cubie, Int)

cubeForest :: Forest Layer

firstCorners :: [LayerSeed]

buildNode :: LayerSeed -> (Layer, [LayerSeed])

combinations

Arguments

:: Bool
set of corners (True) or edges (False)

-> Int
layer size

-> Set Cubie
set of cubies

-> Seq [Cubie]
combinations

Given a set of cubies, produce all possible rotations and permutations.

cubieLocation

Arguments

:: Int
layer number

-> Int
cubie index

-> (Int, Int, Int)
coordinates

Calculate the location of a cubie given the layer number and its position in the layer.

valid :: Layer -> LayerSeed -> Bool

Test whether or not the cubies in the new layer share colours with those on the old layer.