Cube

Description

A 3×3 Rubik's cube is built from three types of Cubies: Centers, Edges, and Corners. These have defined sticker colours and orientations. Such Cubies can be wrapped in a CubieEmbed type by specifying a location in space. Finally, these can be assembled to form a Cube.

This module contains functions for manipulating Cubies, inspecting CubieEmbeds, and printing out the net of a Cube.

Documentation

data Color

Constructors

W
White

Y
Yellow

G
Green

B
Blue

R
Red

O
Orange

data Face

Constructors

Up

Down

Front

Back

Right

Left

data Cubie

A cubie is determined by the colours of its stickers. In addition, this type encodes the orientation of the cubie.

I.e. Corner W G O is the white-green-orange corner in the standard orientation, but Corner O W G is the same corner twisted in clockwise direction.

See corners and edges for the standard orientations.

Constructors

Center !Color

Edge !Color !Color

Corner !Color !Color !Color

Inside

type Cube = Array (Int, Int, Int) Cubie

Assuming the standard color scheme and orientation (see faceToColor), the origin is at the front-up-left corner. The axes are oriented as follows:

x:  Right
y:  Down
z:  Towards the back

data CubieEmbed

This type allows to define a cubie outside the context of an existing cube.

Constructors

CubieEmbed

Fields

cubie :: !Cubie

location :: !(Int, Int, Int)

The index, where the cubie is to be "embedded" in a Cube object

corners :: Set Cubie

The set of all possible Corners in the standard orientation:

{Corner W G O, Corner W O B, Corner W B R, Corner W R G,
 Corner Y O G, Corner Y B O, Corner Y R B, Corner Y G R}

edges :: Set Cubie

The set of all possible Edges in the standard orientation:

{Edge W G, Edge W O, Edge W B, Edge W R,
 Edge G O, Edge B O, Edge B R, Edge G R,
 Edge Y G, Edge Y O, Edge Y B, Edge Y R}

cubiesToCube :: [CubieEmbed] -> Cube

Builds a Cube from a list of 12 edges and 8 corners.

Does not check the input!

isEdge :: Cubie -> Bool

isCorner :: Cubie -> Bool

rotate :: Cubie -> Cubie

Flips an Edge and rotates a Corner clockwise

orient :: Cubie -> Cubie

Returns the cubie in standard orientation

rotNum :: Cubie -> Int

Return the number of time the cubie must be rotated to be oriented

faceToColor :: Face -> Color

Returns the colour of the center cubie of the given face:

Up	Down	Front	Back	Right	Left

W	Y	G	B	R	O

shareColor :: CubieEmbed -> CubieEmbed -> Bool

Checks if there is a face, where the two given cubies have the same colour.

cubieColor :: CubieEmbed -> Face -> Maybe Color

normals :: CubieEmbed -> [(Int, Int, Int)]

Compute the normal vectors that point in the direction the stickers of a given cubie are facing.

normalToFace :: (Int, Int, Int) -> Face

cubeNet :: Cube -> String