Solvability

Description

Not all configurations of a cube can be solved. This module provides functions for checking when a cube can be solved.

Documentation

solvable :: [Cubie] -> Bool

This function assumes that the list is ordered correctly. Below is the order in which the cubies are assumed to appear.

(0,0,0), (1,0,0), (0,1,0), (0,0,1), (2,0,0), (0,2,0), (0,0,2),
(2,1,0), (2,0,1), (0,2,1), (1,2,0), (1,0,2), (0,1,2),
(0,2,2), (2,0,2), (2,2,0), (1,2,2), (2,1,2), (2,2,1), (2,2,2)

This is the order in which the cubies appear, if they are obtained by folding the cubeTree

Note: solvability can be checked simply by computing the parities of the permutation solving the cube, and is thus not too inefficient.

solvableMini :: [Cubie] -> Bool

This is an analogous function for the 2×2 cube.