ConstructForest

Description

The cubeForest is constructed in the following manner:

Start with a list of corners all embedded at the origin.

Let each corner be a root of a tree.

For a given corner, pick three edges and embed them in the three spots adjacent to the origin in such a way that no two adjacent stickers have the same colour.

Take such a triplet of edges to be a child of the root node.

Continue this construction inductively: For each leaf, construct a child node by choosing a set of adjacent pieces, in such a way that no two adjacent stickers have the same colour.

The leaves on level 5 in the forest define all the cubes with the property that no two adjacent stickers have the same colour.

Documentation

type Layer = [CubieEmbed]

cubeForest :: Forest Layer

cubeMiniTree :: Tree Layer

This is an analogous object for the 2×2 cube.