Cube

Description

The 3×3 Rubik's cube is defined as an array of cubies, which have defined sticker colours. This module contains a function for generating a net of a given cube, which can be printed in the terminal.

Documentation

data Color

Possible colours of stickers indicated by a letter.

Constructors

W
White

Y
Yellow

G
Green

B
Blue

R
Red

O
Orange

data Face

Constructors

Up

Down

Front

Back

Right

Left

data Cubie

Constructors

Center !Color

Edge !Color !Color

Corner !Color !Color !Color

Inside

data CubieEmbed

This type allows to define a cubie outside the context of an existing cube.

Constructors

CubieEmbed

Fields

cubie :: !Cubie

location :: !(Int, Int, Int)

The index, where the cubie is to be embedded in a Cube object

type Layer = [CubieEmbed]

Layers are used in the ConstructForest module to build Cubes layer-by-layer

type Cube = Array (Int, Int, Int) Cubie

Assuming the standard color scheme and orientation, where the white face is on top and green is in the front, the origin is at the front-up-left corner. The axes are oriented as in the picture below.

TODO: Add picture

The orientation of the cubie is determined by the order of the colours in the type. See corners and edges for the standard orientations.

rotate :: Cubie -> Cubie

corners :: Set Cubie

The set of all possible corners in the standard orientation:

{Corner W G O, Corner W O B, Corner W B R, Corner W R G, Corner Y O G, Corner Y B O, Corner Y R B, Corner Y G R}

edges :: Set Cubie

The set of all possible edges in the standard orientation:

{Edge W G, Edge W O, Edge W B, Edge W R, Edge G O, Edge B O, Edge B R, Edge G R, Edge Y G, Edge Y O, Edge Y B, Edge Y R}

getFace :: Cube -> Face -> Array (Int, Int) Color

cubeNet :: Cube -> String

cubieColor :: CubieEmbed -> Face -> Maybe Color

emptyNet :: String